// apps/dashboard/convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as _helpers_audit from "../_helpers/audit.js";
import type * as _helpers_exchangeRate from "../_helpers/exchangeRate.js";
import type * as _helpers_permissions from "../_helpers/permissions.js";
import type * as _helpers_shifts from "../_helpers/shifts.js";
import type * as _helpers_stock from "../_helpers/stock.js";
import type * as analytics from "../analytics.js";
import type * as categories from "../categories.js";
import type * as customers from "../customers.js";
import type * as email from "../email.js";
import type * as exchangeRates from "../exchangeRates.js";
import type * as invitations from "../invitations.js";
import type * as members from "../members.js";
import type * as migrations from "../migrations.js";
import type * as notifications from "../notifications.js";
import type * as products from "../products.js";
import type * as registers from "../registers.js";
import type * as returns from "../returns.js";
import type * as sales from "../sales.js";
import type * as shifts from "../shifts.js";
import type * as stockMovements from "../stockMovements.js";
import type * as storeRoles from "../storeRoles.js";
import type * as stores from "../stores.js";
import type * as users from "../users.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

declare const fullApi: ApiFromModules<{
  "_helpers/audit": typeof _helpers_audit;
  "_helpers/exchangeRate": typeof _helpers_exchangeRate;
  "_helpers/permissions": typeof _helpers_permissions;
  "_helpers/shifts": typeof _helpers_shifts;
  "_helpers/stock": typeof _helpers_stock;
  analytics: typeof analytics;
  categories: typeof categories;
  customers: typeof customers;
  email: typeof email;
  exchangeRates: typeof exchangeRates;
  invitations: typeof invitations;
  members: typeof members;
  migrations: typeof migrations;
  notifications: typeof notifications;
  products: typeof products;
  registers: typeof registers;
  returns: typeof returns;
  sales: typeof sales;
  shifts: typeof shifts;
  stockMovements: typeof stockMovements;
  storeRoles: typeof storeRoles;
  stores: typeof stores;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;

export declare const components: {};